move fakerpc to its own package
